Added Login And register Forms
