+++ public/index.css
@@ -1,0 +1,4 @@
+@font-face {
+  font-family: 'Pacifico';
+  src: url(vendors/fonts/Pacifico.ttf);
+}
