+++ public/index.html
@@ -7,14 +7,14 @@
     <meta name="theme-color" content="#000000" />
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
     <link rel="stylesheet" href="vendors/bootstrap/css/bootstrap.min.css" />
-    <link href="vendors/fonts/Pacifico-Regular.ttf" rel="stylesheet" />
+    <link href="index.css" rel="stylesheet" />
     <title>Event | We Build Events</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
 
-    <script src="vendors/bootstrap/js/bootstrap.bundle.min.js"></script>
     <script src="vendors/jquery/jquery.min.js"></script>
+    <script src="vendors/bootstrap/js/bootstrap.bundle.min.js"></script>
   </body>
 </html>
