+++ src/components/Logo.tsx
@@ -1,0 +1,7 @@
+import React, { Component } from 'react';
+
+export class Logo extends Component {
+  render() {
+    return <h3 className="logo flex-grow-1 py-2 px-4">Eventy</h3>;
+  }
+}
