+++ src/containers/Forms.css
@@ -1,0 +1,37 @@
+.fixed {
+  flex-basis: 50%;
+  background-image: url(/vendors/images/concert3.jpg);
+}
+.form {
+  flex-basis: 50%;
+  padding-right: 150px;
+  padding-top: 70px;
+  padding-left: 100px;
+}
+.form-button {
+  border: none;
+  background-color: #ff4d4f;
+  color: white;
+  font-family: sans-serif;
+  min-width: 200px;
+}
+.form-button:hover {
+  background-color: #ff4d4f;
+  color: white;
+}
+.login-form-forgot {
+  float: right;
+}
+.login-form-register {
+  width: 100%;
+  justify-content: center;
+  align-items: center;
+}
+.ant-form-item-label {
+  overflow: hidden;
+  line-height: 39.9999px;
+  white-space: nowrap;
+  text-align: right;
+  vertical-align: middle;
+  display: inline-flex;
+}
