+++ src/containers/LandingPage/components/navbar.tsx
@@ -1,12 +1,13 @@
 import React, { Component } from 'react';
 import { Button } from 'antd';
 import './Navbar.css';
+import { Logo } from '../../../components/Logo';
 export class NavBar extends Component {
   render() {
     return (
-      <nav className="d-flex align-items-center py-2 px-4">
-        <h3 className="logo flex-grow-1">Eventy</h3>
-        <div className="">
+      <nav className="d-flex align-items-center">
+        <Logo />
+        <div className="py-2 px-4">
           <Button className="mr-2 login" type="link">
             LOGIN
           </Button>
