--- src/containers/Registration/RegistrationPage.css
@@ -1,25 +1,0 @@
-.register-form {
-  max-width: 400px;
-  min-width: 350px;
-}
-
-.register-form-button {
-  width: 100%;
-  border: none;
-  background-color: #ff4d4f;
-  color: white;
-  font-family: sans-serif;
-}
-.register-form-button:hover {
-  background-color: #ff4d4f;
-  color: white;
-}
-
-.ant-form-item-label {
-  overflow: hidden;
-  line-height: 39.9999px;
-  white-space: nowrap;
-  text-align: right;
-  vertical-align: middle;
-  display: inline-flex;
-}
